requirements.txt
numpy
pandas
scikit-learn
hdbscan
umap-learn
fastapi
uvicorn
matplotlib
seaborn
pydantic
